updated test files and renamed accordingly
